First crc version that works

+++ board_driver/crc.h
@@ -2,5 +2,5 @@
 #define CRC_H
 
 void crc_init(void);
-uint32_t crc_calculate(uint32_t buffer[], size_t n);
+uint32_t crc_calculate(uint32_t buffer[], uint8_t n);
 #endif
